+++ great_international/cache.py
@@ -18,7 +18,8 @@ class InternationalSectorPageSubscriber(AbstractDatabaseCacheSubscriber):
         great_international.InternationalSubSectorPage,
         great_international.InternationalTopicLandingPage,
         great_international.InternationalCuratedTopicLandingPage,
-        great_international.AboutUkLandingPage
+        great_international.AboutUkLandingPage,
+        great_international.InternationalHomePage
     ]
 
 
@@ -38,6 +39,10 @@ class InternationalHomePageSubscriber(AbstractDatabaseCacheSubscriber):
     model = great_international.InternationalHomePage
     subscriptions = [
         great_international.InternationalArticlePage,
+        invest.InvestInternationalHomePage,
+        capital_invest.InternationalCapitalInvestLandingPage,
+        find_a_supplier.InternationalTradeHomePage,
+        great_international.InternationalSectorPage
     ]
 
 
@@ -113,7 +118,8 @@ class InternationalCapitalInvestLandingPageSubscriber(
     subscriptions = [
         capital_invest.CapitalInvestRegionPage,
         great_international.AboutUkRegionPage,
-        great_international.InternationalGuideLandingPage
+        great_international.InternationalGuideLandingPage,
+        great_international.InternationalHomePage
     ]
 
 
@@ -141,7 +147,8 @@ class InvestInternationalHomePageSubscriber(AbstractDatabaseCacheSubscriber):
     subscriptions = [
         great_international.InternationalSectorPage,
         invest.InvestHighPotentialOpportunityDetailPage,
-        great_international.InternationalGuideLandingPage
+        great_international.InternationalGuideLandingPage,
+        great_international.InternationalHomePage
     ]
 
 
@@ -195,7 +202,8 @@ class InvestRegionPageSubscriber(AbstractDatabaseCacheSubscriber):
 class InternationalTradeHomePageSubscriber(AbstractDatabaseCacheSubscriber):
     model = find_a_supplier.InternationalTradeHomePage
     subscriptions = [
-        great_international.InternationalSectorPage
+        great_international.InternationalSectorPage,
+        great_international.InternationalHomePage
     ]